% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{c3pars}
\alias{c3pars}
\title{\code{c3pars} object constructor}
\usage{
c3pars(t0 = 0, tau1 = 20, tau2 = 0, tau3 = 0, k1 = 0.04, k2 = 0.04,
  k3 = 0, t2ref = 1, squeeze = 0, dummy = FALSE)
}
\arguments{
\item{t0}{Nuclear envelope breakdown time (should be zero)}

\item{tau1}{Timescale for onset of B->P transition}

\item{tau2}{Timescale for delay of P->R transition (zero if no delay)}

\item{tau3}{Timescale for delay of P->B transition (zero if no delay)}

\item{k1}{B->P rate}

\item{k2}{P->R rate}

\item{k3}{R->B rate}

\item{t2ref}{Reference point for time t2, either 1 for "t1" or 0 for "t0"}

\item{dummy}{Logical, if set empty list is returned}
}
\value{
Model parameters object
}
\description{
\code{c3pars} object constructor
}
