% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{cellCount}
\alias{cellCount}
\title{Count colours across cells}
\usage{
cellCount(cells, time, colours)
}
\arguments{
\item{cells}{A matrix with cell colours}

\item{time}{Time vector}

\item{colours}{Factor with colours}
}
\value{
Marginal count
}
\description{
Count colours across cells
}
