% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{ChromCom3}
\alias{ChromCom3}
\title{\code{ChrCom3} object constructor}
\usage{
ChromCom3(pars, time = NULL, cells = NULL, timepars = list(start = -140,
  stop = 90, step = 1), colours = c("B", "P", "R"))
}
\arguments{
\item{pars}{A list of model parameters}

\item{time}{An optional time vector}

\item{cells}{An optional matrix with data}

\item{timepars}{Parameters for the time sequence}

\item{colours}{Colour names}
}
\value{
A \code{ChrCom3} object.
}
\description{
\code{ChromCom3} will create an empty object with model parameters. Optionally, data can be added.
}
