% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{experimentalData}
\alias{experimentalData}
\title{Read experimental data from CSV file}
\usage{
experimentalData(file)
}
\arguments{
\item{file}{File name}
}
\value{
A \code{ChrCom3} object
}
\description{
Read experimental data from CSV file
}
