% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{generateCells}
\alias{generateCells}
\title{Perform simulation}
\usage{
generateCells(chr, nsim = 1000, method = "simulation")
}
\arguments{
\item{chr}{Initial \code{ChrCom3} object with model parameters.}

\item{nsim}{Number of simulations.}

\item{method}{Either "transition" or "simulation"}
}
\value{
A \code{ChrCom3} object with simulation results.
}
\description{
Perform simulation
}
