% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{meltTimelines}
\alias{meltTimelines}
\title{Melt timelines for plotting}
\usage{
meltTimelines(chr, label1 = "L1", label2 = "L2", smooth = FALSE, k = 5)
}
\arguments{
\item{chr}{A \code{ChrCom3} object with data}

\item{label1}{Name of the first variable in the grid}

\item{label2}{Name of the second variable in the grid}

\item{smooth}{If TRUE, smoothing will be applied}

\item{k}{Smoothing window size}
}
\value{
Melted data frame with proportions
}
\description{
Melt timelines for plotting
}
