% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{oeError}
\alias{oeError}
\title{Data-model error}
\usage{
oeError(chr, echr, limits = c(-50, 30))
}
\arguments{
\item{chr}{Model data}

\item{echr}{Experimental data}

\item{limits}{Limiting window for calculations.}
}
\value{
RMS over all three curves
}
\description{
Using test of independence between observed and model counts in the limits window.
}
