% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{plotTimelines}
\alias{plotTimelines}
\title{Plot time lines}
\usage{
plotTimelines(chr, smooth = FALSE, k = 5, expdata = NULL, title = NULL,
  title.size = 10, withpars = FALSE, ...)
}
\arguments{
\item{chr}{A \code{ChrCom3} object with data}

\item{smooth}{If TRUE, smoothing will be applied}

\item{k}{Window size for smoothing}

\item{expdata}{Experimental data to add to the plot (another \code{ChrCom3} object)}

\item{title}{Title of the plot}

\item{title.size}{Font size for the title}

\item{withpars}{Title will be replaced with model parameters}

\item{...}{Other parameters passed to \code{\link{timelinePanel}}}
}
\description{
Plot time lines
}
