% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\docType{data}
\name{simple_theme_grid}
\alias{simple_theme_grid}
\title{Simple theme for plotting}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 57.}
\usage{
simple_theme_grid
}
\description{
Simple theme for plotting
}
\keyword{datasets}
