% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{tcSimulation}
\alias{tcSimulation}
\title{Create a cell using simulations method}
\usage{
tcSimulation(pars, timepars)
}
\arguments{
\item{pars}{Model parameters}

\item{timepars}{Time parameters}
}
\value{
A cell
}
\description{
Create a cell using simulations method
}
