% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{tcTransition}
\alias{tcTransition}
\title{Create a cell using transition method (OBSOLETE)}
\usage{
tcTransition(pars, timepars)
}
\arguments{
\item{pars}{Model parameters}

\item{timepars}{Time parameters}
}
\value{
A cell
}
\description{
Create a cell using transition method (OBSOLETE)
}
