% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{timeIndex}
\alias{timeIndex}
\title{Index of a time point}
\usage{
timeIndex(t, tp, maxn = 10000)
}
\arguments{
\item{t}{Time point in minutes}

\item{tp}{A list of start, stop and step for the timeline}
}
\value{
Integer index in the time vector corresponding to t;
}
\description{
Index of a time point
}
