% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{timelineCell}
\alias{timelineCell}
\title{Generate timeline for one cell}
\usage{
timelineCell(pars, timepars)
}
\arguments{
\item{pars}{Model parameters}

\item{timepars}{A list of start, stop and step for the timeline}
}
\value{
A character vector of colours
}
\description{
Generate timeline for one cell
}
