% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{timelinePanel}
\alias{timelinePanel}
\title{Plot panel with time lines}
\usage{
timelinePanel(m, single = FALSE)
}
\arguments{
\item{m}{Melted data frame}

\item{single}{If true, no facets are applied}
}
\description{
Plot panel with time lines
}
