% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{transitionTimes}
\alias{transitionTimes}
\title{Generate transition times}
\usage{
transitionTimes(pars)
}
\arguments{
\item{pars}{Parameters of the simulation (t1, dt2, k1, k2)}
}
\value{
A list with two transition times
}
\description{
Generate transition times
}
