% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{c3pars}
\alias{c3pars}
\title{\code{c3pars} object constructor}
\usage{
c3pars(tau = -20, dt2 = 0, dt3 = 0, k1 = 0.04, k2 = 0.04, k3 = 0,
  squeeze = 0, dummy = FALSE)
}
\arguments{
\item{tau}{Start time scale}

\item{dt2}{Delay until P->R is possible}

\item{dt3}{Delay until P>B is possible}

\item{k1}{B->P rate}

\item{k2}{P->R rate}

\item{k3}{R->B rate}

\item{dummy}{Logical, if set empty list is returned}
}
\value{
Model parameters object
}
\description{
\code{c3pars} object constructor
}
