% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{c3pars}
\alias{c3pars}
\title{\code{c3pars} object constructor}
\usage{
c3pars(t1 = -30, dt2 = 0, dt3 = 0, k1 = 0.04, k2 = 0.04, k3 = 0,
  dummy = FALSE)
}
\arguments{
\item{dummy}{Logical, if set empty list is returned}
}
\value{
Model parameters object
}
\description{
\code{c3pars} object constructor
}
