% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{fitChr}
\alias{fitChr}
\title{Fit ChromCom data with a model}
\usage{
fitChr(echr, pars, freepars, nsim = 1000, ntry = 10, ncores = 4)
}
\arguments{
\item{echr}{A \code{ChromCom3} object with experimental data}

\item{pars}{A \code{c3pars} object with initial parameters}

\item{freepars}{A character vector with names of free parameters}

\item{nsim}{Number of cells to simulate}

\item{ntry}{Number of tries in search}

\item{ncores}{Number of cores}
}
\value{
A \code{ChromCom3} object with the best-fitting model. "rms" field
  is added. \code{\link{optim}} function is used for minimization with method "L-BFGS-B".
}
\description{
Fits data with a model in which t1, k1, k2 and dt2 are free parameters. If
npar=3, only first three of them are used.
}
