% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{oeError}
\alias{oeError}
\title{Data-model error}
\usage{
oeError(chr, echr, limits = c(-90, 30))
}
\arguments{
\item{chr}{Model data}

\item{echr}{Experimental data}
}
\value{
chi-square
}
\description{
Data-model error
}
