% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{timeIndex}
\alias{timeIndex}
\title{Index of a time point}
\usage{
timeIndex(t, tp)
}
\arguments{
\item{t}{Time point in seconds}

\item{tp}{A list of start, stop and step for the timeline}
}
\value{
Integer index in the time vector corresponding to t; returns n if index outside range
}
\description{
Index of a time point
}
