% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{timelinePanel}
\alias{timelinePanel}
\title{Plot panel with time lines}
\usage{
timelinePanel(m, single = FALSE, xmin = NA, xmax = NA)
}
\arguments{
\item{m}{Melted data frame}

\item{single}{If true, no facets are applied}

\item{xmin}{Lower limit on x-axis}

\item{xmax}{Upper limit on x-axis}
}
\description{
Plot panel with time lines
}
